requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.41
pymysql==1.1.1
email-validator==2.2.0
cryptography==45.0.2
httpx==0.28.1
pytest==8.3.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20